+++ public/js/yith-infinite-scrolling.js
@@ -2,7 +2,7 @@
 	'use strict';
 	$( document ).on( 'yith-infs-scroll-finished', function () {
 		setTimeout( function () {
-			wdevs_gallery_slider_init_swiper();
+			wdevs_gallery_swiper_init_swiper();
 		}, 100 );
 	} );
 } )( jQuery, window, document );