+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 All notable changes to the Wdevs Tax Switch plugin will be documented in this file.
 
+## [1.3.0] - 24-09-2024
+### Updated
+
+- Renamed 'Gallery Slider for WooCommerce' to 'Gallery Swiper for WooCommerce'
+
+## [1.2.0] - 23-08-2024
+### Added
+
+- Declared WooCommerce HPOS compatibility
+
+### Added
+- CHANGELOG.md (this file)
+
 ## [1.1.0] - 17-08-2024
 ### Updated